move handler to internal class
