+++ vaadinfrontend/src/main/java/life/qbic/datamanager/views/project/view/components/ProjectDetailsComponent.java
@@ -15,6 +15,8 @@ import java.util.Objects;
 import java.util.function.Consumer;
 import life.qbic.datamanager.views.layouts.CardLayout;
 import life.qbic.projectmanagement.application.ProjectInformationService;
+import life.qbic.projectmanagement.domain.project.Project;
+import life.qbic.projectmanagement.domain.project.ProjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
@@ -39,45 +41,15 @@ public class ProjectDetailsComponent extends Composite<CardLayout> {
   final TextArea experimentalDesignField = new TextArea();
   final TextArea projectObjective = new TextArea();
 
-  private transient final ProjectDetailsHandler handler;
+  private transient final Handler handler;
 
   public ProjectDetailsComponent(@Autowired ProjectInformationService projectInformationService) {
     Objects.requireNonNull(projectInformationService);
-    setFieldsEditableOnlyOnFocus();
-    attachSubmissionActionOnBlur();
+    this.handler = new Handler(this, projectInformationService);
+    handler.setFieldsEditableOnlyOnFocus();
+    handler.attachSubmissionActionOnBlur();
     initLayout();
     setComponentStyles();
-    this.handler = new ProjectDetailsHandler(this, projectInformationService);
-  }
-
-  private static <T extends Component & HasValue<?, ?> & Focusable<?>> void editableOnFocus(
-      T element) {
-    element.setReadOnly(true);
-    element.addFocusListener(it -> element.setReadOnly(false));
-    element.addBlurListener(it -> element.setReadOnly(true));
-  }
-
-  private static <V, T extends HasValue<?, V> & Focusable<?>> void submitOnBlur(T element,
-      Consumer<V> submitAction) {
-    element.addBlurListener(it -> submitAction.accept(element.getValue()));
-  }
-
-  private void setFieldsEditableOnlyOnFocus() {
-    editableOnFocus(titleField);
-    editableOnFocus(projectObjective);
-    editableOnFocus(experimentalDesignField);
-  }
-
-  private void attachSubmissionActionOnBlur() {
-    submitOnBlur(titleField, value ->
-        //TODO replace with call to application service
-        System.out.println("submitting title = " + value));
-    submitOnBlur(projectObjective, value ->
-        //TODO replace with call to application service
-        System.out.println("submitting objective = " + value));
-    submitOnBlur(experimentalDesignField, value ->
-        //TODO replace with call to application service
-        System.out.println("submitting experimental design description = " + value));
   }
 
   private void initLayout() {
@@ -100,8 +72,77 @@ public class ProjectDetailsComponent extends Composite<CardLayout> {
   public void projectId(String projectId) {
     handler.projectId(projectId);
   }
-  public void setStyles(String... componentStyles){
+
+  public void setStyles(String... componentStyles) {
     getContent().addClassNames(componentStyles);
   }
 
+  /**
+   * Component logic for the {@link ProjectDetailsComponent}
+   *
+   * @since 1.0.0
+   */
+  private final class Handler {
+
+    private final ProjectDetailsComponent component;
+    private final ProjectInformationService projectInformationService;
+
+    private ProjectId selectedProject;
+
+    public Handler(ProjectDetailsComponent component,
+        ProjectInformationService projectInformationService) {
+      this.component = component;
+      this.projectInformationService = projectInformationService;
+    }
+
+    public void projectId(String projectId) {
+      projectInformationService.find(ProjectId.parse(projectId)).ifPresentOrElse(
+          this::loadProjectData,
+          () -> component.titleField.setValue("Not found"));
+    }
+
+    private void loadProjectData(Project project) {
+      this.selectedProject = project.getId();
+      component.titleField.setValue(project.getProjectIntent().projectTitle().title());
+      component.projectObjective.setValue(project.getProjectIntent().objective().value());
+      project.getProjectIntent().experimentalDesign().ifPresentOrElse(
+          experimentalDesignDescription -> component.experimentalDesignField.setValue(
+              experimentalDesignDescription.value()),
+          () -> component.experimentalDesignField.setPlaceholder("No description yet."));
+    }
+
+
+    public void setFieldsEditableOnlyOnFocus() {
+      editableOnFocus(titleField);
+      editableOnFocus(projectObjective);
+      editableOnFocus(experimentalDesignField);
+    }
+
+    public void attachSubmissionActionOnBlur() {
+      ProjectDetailsComponent.Handler.submitOnBlur(titleField, value ->
+          //TODO replace with call to application service
+          System.out.println("project " + selectedProject + ": submitting title = " + value));
+      ProjectDetailsComponent.Handler.submitOnBlur(projectObjective, value ->
+          //TODO replace with call to application service
+          System.out.println("project " + selectedProject + ": submitting objective = " + value));
+      ProjectDetailsComponent.Handler.submitOnBlur(experimentalDesignField, value ->
+          //TODO replace with call to application service
+          System.out.println(
+              "project " + selectedProject + ": submitting experimental design description = "
+                  + value));
+    }
+
+    private static <T extends Component & HasValue<?, ?> & Focusable<?>> void editableOnFocus(
+        T element) {
+      element.setReadOnly(true);
+      element.addFocusListener(it -> element.setReadOnly(false));
+      element.addBlurListener(it -> element.setReadOnly(true));
+    }
+
+    private static <V, T extends HasValue<?, V> & Focusable<?>> void submitOnBlur(T element,
+        Consumer<V> submitAction) {
+      element.addBlurListener(it -> submitAction.accept(element.getValue()));
+    }
+
+  }
 }
